added a divide view over two lists
